refactor: Improve rank comparison formatting and type consistency
- Change recomposition naming from "Recomposition1" to "Recomposition.1" format
- Add explicit dtype=int to rank value arrays in tests
- Remove trailing whitespace and clean up formatting

+++ tests/ranksrev/test_rank_transitivity_check.py
@@ -56,8 +56,7 @@ def test_TransitivityChecker():
     dm = skc.datasets.load_simple_stock_selection()
 
     result = dec.evaluate(dm)
-
-    # expected = RanksComparator(
+    
     expected = RanksComparator(
         [
             (
@@ -65,7 +64,7 @@ def test_TransitivityChecker():
                 RankResult(
                     method="WeightedSumModel",
                     alternatives=["PE", "JN", "AA", "FX", "MM", "GN"],
-                    values=np.array([3, 4, 2, 1, 5, 5]),
+                    values=np.array([3, 4, 2, 1, 5, 5], dtype=int),
                     extra={},
                 ),
             ),
@@ -74,7 +73,7 @@ def test_TransitivityChecker():
                 RankResult(
                     method="WeightedSumModel",
                     alternatives=["PE", "JN", "AA", "FX", "MM", "GN"],
-                    values=np.array([3, 4, 2, 1, 5, 5]),
+                    values=np.array([3, 4, 2, 1, 5, 5], dtype=int),
                     extra={},
                 ),
             ),
